agenda_api/requirements/common.txt
# Core Stuff
# -------------------------------------
Django==3.1.8

# Configuration
# -------------------------------------
argon2-cffi==20.1.0
django-environ==0.4.5
# For Django 3.x
-e git+git://github.com/CuriousLearner/django-sites@upgrade-django-3.x#egg=django_sites
python-dotenv==0.17.0
django-cors-headers==3.7.0

# Extensions
# -------------------------------------
pytz==2021.1

# Models
# -------------------------------------
psycopg2-binary==2.8.6

Pillow==8.2.0
django-extensions==3.1.3
django-uuid-upload-path==1.0.0

# REST APIs
# -------------------------------------
djangorestframework==3.12.4
django-rest-swagger==2.2.0

# LOGGING
# -------------------------------------
django-log-request-id==1.6.0

# Auth Stuff
# -------------------------------------
PyJWT==2.0.1
django-mail-templated==2.6.5

# Generating DB Schema
# --------------------------------------
pydot==1.4.2
pyparsing==2.4.7
